Move Map enum functions to separate file
