+++ src/GUI/UI/UIPopup.cpp
@@ -23,6 +23,7 @@
 #include "Engine/Tables/CharacterFrameTable.h"
 #include "Engine/Spells/Spells.h"
 #include "Engine/Party.h"
+#include "Engine/MapEnumFunctions.h"
 #include "Engine/Time/Timer.h"
 #include "Engine/Conditions.h"
